+++ system/domains/billing/src/main/kotlin/ru/upg/ates/billing/table/TaskTable.kt
@@ -5,7 +5,7 @@ import ru.upg.ates.table.AtesTable
 object TaskTable : AtesTable("task") {
     val userId = long("user_id").references(UserTable.id)
     val name = text("name")
-    val assignPrice = integer("assign_price")
-    val finishPrice = integer("finish_price")
+    val assignPrice = long("assign_price")
+    val finishPrice = long("finish_price")
     val finished = bool("finished").default(false)
 }
